multi-agent-mcp/pyproject.toml
[project]
name = "multi-agent-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "mcp[cli]>=1.12.2",
    "requests>=2.32.4",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]
